requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.9
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
deprecation==2.1.0
exceptiongroup==1.3.0
frozenlist==1.6.2
google-auth==2.40.3
google-genai==1.19.0
gotrue==2.12.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
multidict==6.4.4
numpy>=1.21.0,<2.0.0
opencv-python>=4.8.0
packaging>=25.0
pluggy>=1.6.0
postgrest>=1.0.2
propcache>=0.3.1
pyasn1>=0.6.1
pyasn1_modules>=0.4.2
pydantic>=2.11.5
pydantic_core>=2.33.2
Pygments>=2.19.1
PyJWT>=2.10.1
pytest>=8.4.0
pytest-mock>=3.14.1
python-dateutil>=2.9.0
python-dotenv>=1.1.0
realtime>=2.4.3
requests>=2.32.3
rsa>=4.9.1
six>=1.17.0
sniffio>=1.3.1
storage3>=0.11.3
StrEnum>=0.4.15
supabase>=2.15.2
supafunc>=0.9.4
tomli>=2.2.1
typing-inspection>=0.4.1
typing_extensions>=4.14.0
urllib3>=2.4.0
websockets>=14.2
yarl>=1.20.0
google-api-python-client>=2.150.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
tqdm>=4.67.0
pytz>=2025.1
pandas>=1.5.0
